backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pyyaml>=6.0
httpx>=0.25.0
psdl-lang>=0.2.0,<0.3.0  # Compatible with PyPI v0.2.x API
